requirements.txt
# Image Captioning dengan CNN-LSTM & FastText
# ============================================
# Dependencies untuk project Image Captioning Bahasa Indonesia

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
scikit-learn>=1.3.0

# NLP & Embeddings
gensim>=4.3.0
nltk>=3.8.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0

# Dataset Management
kagglehub>=0.3.0

# Optional (untuk visualisasi)
seaborn>=0.12.0
jupyter>=1.0.0
